\name{classifyType}
\alias{classifyType}
\title{classifyType}
\description{Classifies attributes according to R data types}
\usage{classifyType(theBaseType)}
\arguments{
  \item{theBaseType}{The xs: namespace-specific definition name of the value}
}



\author{Lucy Bastin and Ankita Goyal}





