\name{isUncertaintyType}
\alias{isUncertaintyType}
\title{isUncertaintyType}
\description{Gives the name of the base type for the given definition of the data type}
\usage{isUncertaintyType(theDoc, typeDef, namespaces, logLevel = 1)}
\arguments{
  \item{theDoc}{All or part of the XML Schema document containing the given type definition}
  \item{typeDef}{The given type definition}
  \item{namespaces}{The namespaces present in the document}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}



\author{Lucy Bastin and Ankita Goyal}





