\name{xmlschemaHelper-package} \alias{xmlschemaHelper-package} \docType{package} \title{Tools for reading XML Schema within R} \description{Reads simple XML schema and returns details of user-specified elements} \details{ \tabular{ll}{Package: \tab xmlschemaHelper\cr Title: \tab Tools for reading XML Schema within R\cr Version: \tab 1.0.1\cr Date: \tab 2013-01-22\cr Depends: \tab R (>= 2.5.0), XML\cr Author: \tab Lucy Bastin and Ankita Goyal\cr Maintainer: \tab Lucy Bastin <l.bastin@aston.ac.uk>\cr License: \tab GPL (>= 2)\cr Repository: \tab CRAN\cr} } \author{Lucy Bastin and Ankita Goyal}  \keyword{ package }   